+++ src/main/java/seedu/address/logic/commands/AddResumeCommand.java
@@ -1,0 +1,40 @@
+package seedu.address.logic.commands;
+
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+import seedu.address.model.item.Item;
+import seedu.address.model.item.Resume;
+
+import static java.util.Objects.requireNonNull;
+
+public class AddResumeCommand extends AddCommand {
+
+    private final Resume toAdd;
+
+    /**
+     * Creates an AddCommand to add the specified {@code Person}
+     */
+    public AddResumeCommand(Resume item) {
+        requireNonNull(item);
+        toAdd = item;
+    }
+
+    @Override
+    public CommandResult execute(Model model) throws CommandException {
+        requireNonNull(model);
+
+        if (model.hasItem(toAdd)) {
+            throw new CommandException(MESSAGE_DUPLICATE_PERSON);
+        }
+
+//        model.addItem(toAdd);
+        return new CommandResult(String.format(MESSAGE_SUCCESS, toAdd.getType().getFullType(), toAdd));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof AddResumeCommand // instanceof handles nulls
+                && toAdd.equals(((AddResumeCommand) other).toAdd));
+    }
+}
